+++ RotateToolbar/Properties/AssemblyInfo.cs
@@ -3,5 +3,5 @@ using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("RotateToolbar")]
 [assembly: Guid("3ed4784c-6ea7-4dbe-8476-4e1fa972fc3d")]
-[assembly: AssemblyVersion("1.3.0.0")]
-[assembly: AssemblyFileVersion("1.3.0.0")]
+[assembly: AssemblyVersion("1.2.1.0")]
+[assembly: AssemblyFileVersion("1.2.1.0")]